Add services for Excel file operations, validation, merging, and console interaction
- Implemented ExcelService for handling Excel file operations including sheet existence checks and column mappings.
- Created IExcelService interface to define contract for Excel operations.
- Developed IValidationService and ValidationService for validating RVTools Excel files against required sheets and mandatory columns.
- Introduced IMergeService and MergeService to facilitate merging of multiple RVTools Excel files with validation and anonymization options.
- Added IAnonymizationService interface for anonymizing data in Excel files.
- Implemented ICommandLineParser for parsing command line arguments.
- Created IConsoleService and SpectreConsoleService for improved console output and progress reporting.
- Established interfaces for various services to enhance testability and maintainability.
